+++ tests/api/test_checkout.py
@@ -279,10 +279,7 @@ def test_checkout_available_payment_gateways(
 
 
 def test_checkout_available_shipping_methods(
-        api_client, checkout_with_item, address, shipping_zone):
-    checkout_with_item.shipping_address = address
-    checkout_with_item.save()
-
+        api_client, checkout_with_item_and_address, shipping_zone):
     query = """
     query getCheckout($token: UUID!) {
         checkout(token: $token) {
@@ -292,7 +289,7 @@ def test_checkout_available_shipping_methods(
         }
     }
     """
-    variables = {'token': checkout_with_item.token}
+    variables = {'token': checkout_with_item_and_address.token}
     response = api_client.post_graphql(query, variables)
     content = get_graphql_content(response)
     data = content['data']['checkout']
@@ -441,7 +438,7 @@ def test_checkout_lines_add_check_lines_quantity(
     assert data['errors'][0]['field'] == 'quantity'
 
 
-def test_checkout_lines_invalid_variant_id(user_api_client, checkout, variant):
+def test_checkout_lines_add_invalid_variant_id(user_api_client, checkout, variant):
     variant_id = graphene.Node.to_global_id('ProductVariant', variant.pk)
     invalid_variant_id = 'InvalidId'
     checkout_id = graphene.Node.to_global_id('Checkout', checkout.pk)
@@ -462,6 +459,27 @@ def test_checkout_lines_invalid_variant_id(user_api_client, checkout, variant):
         'Could not resolve to a node with the global id list of \'%s\'.')
     assert data['errors'][0]['message'] == error_msg % [invalid_variant_id]
     assert data['errors'][0]['field'] == 'variantId'
+
+
+def test_checkout_lines_add_clean_shipping(
+        user_api_client, checkout_with_invalid_shipping_method, variant):
+    checkout = checkout_with_invalid_shipping_method
+    variant_id = graphene.Node.to_global_id('ProductVariant', variant.pk)
+    checkout_id = graphene.Node.to_global_id('Checkout', checkout.pk)
+
+    variables = {
+        'checkoutId': checkout_id,
+        'lines': [{
+            'variantId': variant_id,
+            'quantity': 1}]}
+    response = user_api_client.post_graphql(
+        MUTATION_CHECKOUT_LINES_ADD, variables)
+    content = get_graphql_content(response)
+    data = content['data']['checkoutLinesAdd']
+    assert not data['errors']
+    checkout.refresh_from_db()
+    assert checkout.shipping_address is not None
+    assert checkout.shipping_method is None
 
 
 MUTATION_CHECKOUT_LINES_UPDATE = """
@@ -547,6 +565,30 @@ def test_checkout_lines_update_check_lines_quantity(
     assert data['errors'][0]['field'] == 'quantity'
 
 
+def test_checkout_lines_update_clean_shipping(
+        user_api_client, checkout_with_invalid_shipping_method):
+    checkout = checkout_with_invalid_shipping_method
+    line = checkout.lines.first()
+    variant = line.variant
+    variant_id = graphene.Node.to_global_id('ProductVariant', variant.pk)
+    checkout_id = graphene.Node.to_global_id('Checkout', checkout.pk)
+
+    variables = {
+        'checkoutId': checkout_id,
+        'lines': [{
+            'variantId': variant_id,
+            'quantity': 1}]}
+    response = user_api_client.post_graphql(
+        MUTATION_CHECKOUT_LINES_UPDATE, variables)
+    content = get_graphql_content(response)
+
+    data = content['data']['checkoutLinesUpdate']
+    assert not data['errors']
+    checkout.refresh_from_db()
+    assert checkout.shipping_address is not None
+    assert checkout.shipping_method is None
+
+
 MUTATION_CHECKOUT_LINES_DELETE = """
     mutation checkoutLineDelete($checkoutId: ID!, $lineId: ID!) {
         checkoutLineDelete(checkoutId: $checkoutId, lineId: $lineId) {
@@ -586,6 +628,25 @@ def test_checkout_line_delete(user_api_client, checkout_with_item):
     assert not data['errors']
     checkout.refresh_from_db()
     assert checkout.lines.count() == 0
+
+
+def test_checkout_line_delete_clean_shipping(
+        user_api_client, checkout_with_invalid_shipping_method):
+    checkout = checkout_with_invalid_shipping_method
+    line = checkout.lines.first()
+    checkout_id = graphene.Node.to_global_id('Checkout', checkout.pk)
+    line_id = graphene.Node.to_global_id('CheckoutLine', line.pk)
+
+    variables = {'checkoutId': checkout_id, 'lineId': line_id}
+    response = user_api_client.post_graphql(
+        MUTATION_CHECKOUT_LINES_DELETE, variables)
+    content = get_graphql_content(response)
+
+    data = content['data']['checkoutLineDelete']
+    assert not data['errors']
+    checkout.refresh_from_db()
+    assert checkout.shipping_address is None
+    assert checkout.shipping_method is None
 
 
 def test_checkout_customer_attach(
@@ -819,6 +880,49 @@ def test_checkout_shipping_address_update_invalid_argument(
     assert data['errors'][0]['message'] == (
         'One and only one shipping address or shipping address ID '
         'should be provided.')
+
+
+def test_checkout_shipping_address_update_shipping_not_required(
+        user_api_client, checkout, graphql_address_data):
+    assert checkout.shipping_address is None
+    checkout_id = graphene.Node.to_global_id('Checkout', checkout.pk)
+
+    shipping_address = graphql_address_data
+    variables = {
+        'checkoutId': checkout_id,
+        'shippingAddress': shipping_address}
+
+    response = user_api_client.post_graphql(
+        MUTATION_CHECKOUT_SHIPPING_ADDRESS_UPDATE, variables)
+    content = get_graphql_content(response)
+    data = content['data']['checkoutShippingAddressUpdate']
+    assert len(data['errors']) == 1
+    assert data['errors'][0]['message'] == \
+        'This checkout does not requires shipping.'
+    checkout.refresh_from_db()
+    assert checkout.shipping_address is None
+
+
+def test_checkout_shipping_address_update_keep_previous_address_when_error(
+        user_api_client, checkout_with_item_and_address, graphql_address_data):
+    checkout = checkout_with_item_and_address
+    checkout_id = graphene.Node.to_global_id('Checkout', checkout.pk)
+
+    checkout.lines.all().delete()
+    shipping_address = graphql_address_data
+    variables = {
+        'checkoutId': checkout_id,
+        'shippingAddress': shipping_address}
+
+    response = user_api_client.post_graphql(
+        MUTATION_CHECKOUT_SHIPPING_ADDRESS_UPDATE, variables)
+    content = get_graphql_content(response)
+    data = content['data']['checkoutShippingAddressUpdate']
+    assert len(data['errors']) == 1
+    assert data['errors'][0]['message'] == \
+        'This checkout does not requires shipping.'
+    checkout.refresh_from_db()
+    assert checkout.shipping_address is not None
 
 
 MUTATION_CHECKOUT_BILLING_ADDRESS_UPDATE = """
@@ -1183,11 +1287,7 @@ def test_checkout_prices(user_api_client, checkout_with_item):
         checkout_with_item.get_subtotal().gross.amount)
 
 
-@patch('saleor.graphql.checkout.mutations.clean_shipping_method')
-def test_checkout_shipping_method_update(
-        mock_clean_shipping, staff_api_client,
-        shipping_method, checkout_with_item):
-    query = """
+MUTATION_CHECKOUT_SHIPPING_METHOD_UPDATE = """
     mutation checkoutShippingMethodUpdate(
             $checkoutId:ID!, $shippingMethodId:ID!){
         checkoutShippingMethodUpdate(
@@ -1202,12 +1302,19 @@ def test_checkout_shipping_method_update(
         }
     }
     """
+
+
+@patch('saleor.graphql.checkout.mutations.clean_shipping_method')
+def test_checkout_shipping_method_update(
+        mock_clean_shipping, staff_api_client,
+        checkout_with_item, shipping_method):
     checkout = checkout_with_item
     checkout_id = graphene.Node.to_global_id('Checkout', checkout.pk)
     method_id = graphene.Node.to_global_id(
         'ShippingMethod', shipping_method.id)
     variables = {'checkoutId': checkout_id, 'shippingMethodId': method_id}
-    response = staff_api_client.post_graphql(query, variables)
+    response = staff_api_client.post_graphql(
+        MUTATION_CHECKOUT_SHIPPING_METHOD_UPDATE, variables)
     content = get_graphql_content(response)
     data = content['data']['checkoutShippingMethodUpdate']
     assert not data['errors']
@@ -1216,8 +1323,94 @@ def test_checkout_shipping_method_update(
     checkout.refresh_from_db()
     assert checkout.shipping_method == shipping_method
     mock_clean_shipping.assert_called_once_with(
-        checkout=checkout, method=shipping_method, discounts=ANY, taxes=ANY,
+        checkout=checkout, shipping_method=shipping_method, discounts=ANY, taxes=ANY,
         remove=False)
+
+
+def test_checkout_shipping_method_update_shipping_not_required(
+        staff_api_client, checkout, shipping_method):
+    checkout_id = graphene.Node.to_global_id('Checkout', checkout.pk)
+    method_id = graphene.Node.to_global_id(
+        'ShippingMethod', shipping_method.id)
+    variables = {'checkoutId': checkout_id, 'shippingMethodId': method_id}
+    response = staff_api_client.post_graphql(
+        MUTATION_CHECKOUT_SHIPPING_METHOD_UPDATE, variables)
+    content = get_graphql_content(response)
+    data = content['data']['checkoutShippingMethodUpdate']
+    assert len(data['errors']) == 1
+    assert data['errors'][0]['message'] == (
+        'Cannot choose a shipping method for a '
+        'checkout without the shipping address.')
+    checkout.refresh_from_db()
+    assert checkout.shipping_method is None
+
+
+def test_checkout_shipping_method_update_no_shipping_address(
+        staff_api_client, checkout_with_item, shipping_method):
+    checkout = checkout_with_item
+    checkout_id = graphene.Node.to_global_id('Checkout', checkout.pk)
+    method_id = graphene.Node.to_global_id(
+        'ShippingMethod', shipping_method.id)
+    variables = {'checkoutId': checkout_id, 'shippingMethodId': method_id}
+    response = staff_api_client.post_graphql(
+        MUTATION_CHECKOUT_SHIPPING_METHOD_UPDATE, variables)
+    content = get_graphql_content(response)
+    data = content['data']['checkoutShippingMethodUpdate']
+    assert len(data['errors']) == 1
+    assert data['errors'][0]['message'] == (
+        'Cannot choose a shipping method for a '
+        'checkout without the shipping address.')
+    checkout.refresh_from_db()
+    assert checkout.shipping_method is None
+
+
+def test_checkout_shipping_method_update_invalid_method(
+        staff_api_client, checkout_with_item_and_address,
+        shipping_zone_without_countries):
+    checkout = checkout_with_item_and_address
+    assert checkout.shipping_address is not None
+
+    shipping_method = shipping_zone_without_countries.shipping_methods.first()
+    checkout_id = graphene.Node.to_global_id('Checkout', checkout.pk)
+    method_id = graphene.Node.to_global_id(
+        'ShippingMethod', shipping_method.id)
+    variables = {'checkoutId': checkout_id, 'shippingMethodId': method_id}
+    response = staff_api_client.post_graphql(
+        MUTATION_CHECKOUT_SHIPPING_METHOD_UPDATE, variables)
+    content = get_graphql_content(response)
+    data = content['data']['checkoutShippingMethodUpdate']
+    assert len(data['errors']) == 1
+    assert data['errors'][0]['message'] == \
+        'Shipping method cannot be used with this checkout.'
+    checkout.refresh_from_db()
+    assert checkout.shipping_address is not None
+    assert checkout.shipping_method is None
+
+
+def test_checkout_shipping_method_update_keep_previous_method_when_error(
+        staff_api_client, checkout_with_item_and_address,
+        shipping_method, shipping_zone_without_countries):
+    checkout = checkout_with_item_and_address
+    checkout.shipping_method = shipping_method
+    checkout.save()
+    assert checkout.shipping_address is not None
+    assert checkout.shipping_method is not None
+
+    shipping_method = shipping_zone_without_countries.shipping_methods.first()
+    checkout_id = graphene.Node.to_global_id('Checkout', checkout.pk)
+    method_id = graphene.Node.to_global_id(
+        'ShippingMethod', shipping_method.id)
+    variables = {'checkoutId': checkout_id, 'shippingMethodId': method_id}
+    response = staff_api_client.post_graphql(
+        MUTATION_CHECKOUT_SHIPPING_METHOD_UPDATE, variables)
+    content = get_graphql_content(response)
+    data = content['data']['checkoutShippingMethodUpdate']
+    assert len(data['errors']) == 1
+    assert data['errors'][0]['message'] == \
+        'Shipping method cannot be used with this checkout.'
+    checkout.refresh_from_db()
+    assert checkout.shipping_address is not None
+    assert checkout.shipping_method is not None
 
 
 def test_query_checkout_line(checkout_with_item, user_api_client):
@@ -1440,7 +1633,7 @@ MUTATION_CHECKOUT_UPDATE_VOUCHER = """
             }
         }
     }
-"""
+    """
 
 
 def _mutate_checkout_update_voucher(client, variables):