+++ CHANGELOG.md
@@ -3,7 +3,6 @@
 All notable, unreleased changes to this project will be documented in this file. For the released changes, please visit the [Releases](https://github.com/mirumee/saleor/releases) page.
 
 ## [Unreleased]
-
 - Dropped Python 3.5 support - #4028 by @korycins
 - Refactor error handling in mutations #3891 by @maarcingebala @akjanik
 - Use only_fields instead of exclude_fields in gql api - #3940 by @michaljelonek
@@ -44,6 +43,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Use dataclasses as a structure for payment's gateways - #4028 by @korycins
 - Fix the product listing not showing in the voucher when there were products selected - #4062 by @NyanKiyoshi
 - Fix crash in Dashboard 1.0 when updating an order address's phone number - #4061 by @NyanKiyoshi
+- Improvements on shipping validation - #3812 by @jxltom
 
 
 ## 2.5.0